Restructure key dispatch into peer controllers
