+++ src/composition/hangul-controller.ts
@@ -1,0 +1,395 @@
+import { isModifierKey, isModifierKeyActive, KeyCode, keyMap } from "../keyboard/korean-keyboard-map";
+import { KeyBinding, defaultToggleKeyBindingForPlatform, isModifierOnlyBinding } from "../keyboard/key-binding";
+import { cancelEvent, isKimeEvent } from "../messaging/dom-events";
+import { CompositionAdapter } from "./composition-adapters/composition-adapter";
+import { HangulCompositor } from "./hangul-compositor";
+
+/**
+ * Controls Hangul composition for a given editor adapter.
+ */
+export class HangulController {
+    private _isActive = false;
+    private compositor = new HangulCompositor();
+    private changeListeners: (() => void)[] = [];
+
+    // The most recent modifier key pressed (any side). Combined with the event's live
+    // modifier flags, this lets isToggleKeyHeld tell the configured toggle key apart
+    // from its same-type sibling on the other side; `event.code` is the key pressed
+    // now, and a flag like `metaKey` does not say which side is down.
+    private lastModifierKey?: KeyCode = undefined;
+
+    // The configured Han/Yong toggle key. When it is a modifier-only key (e.g. Right
+    // Alt, or Right Command on macOS) and held down, it acts as the IME key rather
+    // than a Ctrl/Cmd/Alt modifier. Defaults to the platform default; the content
+    // script pushes the user's actual binding via setToggleKeyBinding.
+    private toggleKeyBinding: KeyBinding | null = defaultToggleKeyBindingForPlatform();
+
+    constructor(private readonly compositionAdapter: CompositionAdapter) {}
+
+    get isActive() {
+        return this._isActive;
+    }
+
+    onEntry(listener: () => void) {
+        this.changeListeners.push(listener);
+    }
+
+    activate() {
+        this._isActive = true;
+    }
+
+    deactivate() {
+        this.endComposition();
+        this._isActive = false;
+    }
+
+    dispose() {
+        try {
+            this.flushComposition();
+        } finally {
+            this._isActive = false;
+            this.changeListeners = [];
+        }
+    }
+
+    getCompositionFeatures() {
+        return this.compositionAdapter.getSupportedMethods();
+    }
+
+    /**
+     * Set the configured Han/Yong toggle key (or null when the user turned it off).
+     * Pushed in by the content script so the controller knows which physical modifier
+     * key currently acts as the IME key rather than a Ctrl/Cmd/Alt modifier.
+     */
+    setToggleKeyBinding(binding: KeyBinding | null) {
+        this.toggleKeyBinding = binding;
+    }
+
+    handleKey(event: KeyboardEvent): void {
+        // Do not process events that were generated by KIME.
+        if (isKimeEvent(event)) {
+            return;
+        }
+
+        const code = event.code as KeyCode;
+
+        // Record the most recent modifier key (any side) and stop. Modifier keys
+        // are not text. Tracking which one lets isToggleKeyHeld tell the configured
+        // toggle key apart from its same-type sibling on the other side.
+        if (isModifierKey(code)) {
+            this.lastModifierKey = code;
+            return;
+        }
+
+        if (!this._isActive) {
+            // The on-screen keyboard may have a composition in progress even
+            // though the physical keyboard is inactive (Hangul typing disabled).
+            // A physical keystroke moves past it, so commit/clear it first;
+            // otherwise the next OSK jamo would attach to a stale block.
+            this.flushComposition();
+
+            if (this.isToggleKeyHeld(event)) {
+                // The configured toggle key (e.g. Right Alt, or Right Command on
+                // macOS) is held: it is the IME key, not a modifier, so insert the
+                // character instead of letting the OS treat Alt/Cmd+key as a
+                // shortcut or menu trigger.
+                this.compositionAdapter.inputCharacter(event.key, code);
+                cancelEvent(event);
+                return;
+            }
+
+            return;
+        }
+
+        // Shift+Backspace lifts the previous character back into composition.
+        // Normally already handled in the capture phase (see handleCaptureKey);
+        // this is the fallback for when the capture listener did not run (a detached
+        // element, as in unit tests).
+        const reqPrevCharComposition = !this.compositor.isCompositing() && event.shiftKey && code === KeyCode.Backspace;
+
+        if (reqPrevCharComposition && this.handlePreviousCharComposition(event)) {
+            return;
+        }
+
+        const key = keyMap[code];
+        // event.key is a single character for printable keys, but a multi-character name
+        // for functional keys ("Tab", "Enter", "CapsLock", ...). keyMap lists those so the
+        // on-screen keyboard can label them, but they must never be inserted as text, so
+        // only single-character keys are treated as input below.
+        const isCharacterKey = event.key.length === 1;
+
+        // Normally already handled in the capture phase (see handleCaptureKey);
+        // this remains as a fallback for cases where the capture listener did not
+        // run (e.g. an element detached from the document, as in unit tests).
+        if (code === KeyCode.Backspace && this.compositor.isCompositing()) {
+            this.handleComposingBackspace(event);
+            return;
+        }
+
+        // Do not interfere with shortcuts, functional keys (Tab, Enter, ...), or keys we
+        // do not understand. Commit any in-progress composition and let the browser
+        // handle them natively.
+        if (!key || this.isShortcutChord(event) || !isCharacterKey) {
+            this.endComposition();
+            return;
+        }
+
+        if (!key.jamo) {
+            if (!this.compositor.isCompositing()) {
+                return;
+            }
+
+            this.endComposition();
+
+            // CKEditor throws errors and the character is not inputted unless we add this timeout.
+            window.setTimeout(() => {
+                this.compositionAdapter.inputCharacter(event.key, code);
+            }, 0);
+
+            event.preventDefault();
+            event.stopImmediatePropagation();
+            event.stopPropagation();
+            return;
+        }
+
+        // Fallback for when the capture guard did not run (no selection to replace,
+        // or a detached element as in unit tests). On a connected element with a
+        // selection the guard already handled this jamo and stopped the event.
+        this.handleJamoKey(event);
+    }
+
+    /**
+     * Capture-phase keydown handling, called by KeyListener on window so it runs before
+     * any page handler. It intercepts the two cases where a rich editor (Word for the
+     * Web) acts on a key in its own capture-phase handler before our bubble handler can
+     * run:
+     *
+     *   1. Backspace while composing: the editor deletes the whole composing block.
+     *   2. The composition-starting jamo while text is selected: the editor replaces
+     *      the selection with the key's literal character ("type over selection").
+     *
+     * Everything else is left to the bubble-phase handler.
+     */
+    handleCaptureKey(event: KeyboardEvent): void {
+        if (isKimeEvent(event)) {
+            return;
+        }
+
+        // The composition-capture handling below only applies while Hangul typing is active.
+        if (!this._isActive) {
+            return;
+        }
+
+        const code = event.code as KeyCode;
+
+        if (code === KeyCode.Backspace) {
+            if (this.compositor.isCompositing()) {
+                this.handleComposingBackspace(event);
+            } else if (event.shiftKey) {
+                // Shift+Backspace must run here, ahead of the editor deleting the
+                // previous character itself. Otherwise getPreviousCharacter() returns
+                // the character before the one the editor just removed, and we lift
+                // the wrong one into composition.
+                this.handlePreviousCharComposition(event);
+            }
+            // A plain (non-shift) Backspace while not composing is left to the editor.
+            return;
+        }
+
+        // Begin composition ahead of the editor's "type over selection" only when
+        // there is a real, non-collapsed document selection to replace. The
+        // no-selection path and plain <input> elements (whose selection is not a
+        // document selection) stay on the bubble handler, unchanged.
+        if (!this.compositor.isCompositing() && this.isJamoKey(event) && this.hasNonCollapsedSelection()) {
+            this.handleJamoKey(event);
+        }
+    }
+
+    // Commit and clear any in-progress composition. Runs when the controller is
+    // active, or whenever a composition is in progress even though inactive. The
+    // on-screen keyboard can drive composition while the physical keyboard is
+    // inactive (Hangul typing disabled), so a focus/caret change or a physical
+    // keystroke must still flush it, or the next OSK jamo attaches to a stale block.
+    flushComposition() {
+        if (!this._isActive && !this.compositor.isCompositing()) {
+            return;
+        }
+
+        this.compositor.reset();
+        this.compositionAdapter.blur();
+    }
+
+    // Commit and clear an in-progress composition, if there is one. KeyListener uses
+    // this before Hanja conversion so Hanja can work only from committed text.
+    endComposition() {
+        if (!this.compositor.isCompositing()) {
+            return;
+        }
+
+        this.compositionAdapter.endComposition(this.compositor.getCurrentChar());
+        this.compositor.reset();
+    }
+
+    /**
+     * Add a non-Hangul character to the composition adapter.
+     * This will end any current composition.
+     */
+    addCharacter(char: string, keyCode: KeyCode) {
+        this.endComposition();
+        this.compositionAdapter.inputCharacter(char, keyCode);
+    }
+
+    handleBackspace() {
+        if (this.compositor.isCompositing()) {
+            const block = this.compositor.removeLastJamo();
+            if (block) {
+                this.compositionAdapter.updateComposition(block, KeyCode.Backspace);
+            } else {
+                this.compositionAdapter.endComposition("");
+            }
+
+            this.notifyOnEntry();
+        } else {
+            this.compositionAdapter.deleteContentBackwards();
+        }
+    }
+
+    addJamo(jamo: string, keyCode: KeyCode) {
+        const compositionProgress = this.compositor.addJamo(jamo);
+
+        if ("completed" in compositionProgress) {
+            this.compositionAdapter.endComposition(compositionProgress.completed);
+        }
+
+        if ("started" in compositionProgress) {
+            this.compositionAdapter.beginComposition(compositionProgress.started, keyCode);
+        }
+
+        if ("updated" in compositionProgress) {
+            this.compositionAdapter.updateComposition(compositionProgress.updated, keyCode);
+        }
+
+        this.notifyOnEntry();
+    }
+
+    private notifyOnEntry() {
+        this.changeListeners.forEach((listener) => {
+            try {
+                listener();
+            } catch (e) {
+                console.error(e);
+            }
+        });
+    }
+
+    private isJamoKey(event: KeyboardEvent): boolean {
+        const code = event.code as KeyCode;
+        return !this.isShortcutChord(event) && event.key.length === 1 && !!keyMap[code]?.jamo;
+    }
+
+    /**
+     * Whether the configured Han/Yong toggle key is a modifier-only key that is
+     * currently held. When true, that key is acting as the IME key, not as a
+     * Ctrl/Cmd/Alt modifier, so a letter pressed with it should be composed/typed
+     * rather than treated as a browser shortcut. We require both that the toggle key
+     * was the last modifier pressed (so it is that side, not its sibling) and that its
+     * modifier flag is still down (so it was not already released).
+     */
+    private isToggleKeyHeld(event: KeyboardEvent): boolean {
+        const binding = this.toggleKeyBinding;
+        if (!binding || !isModifierOnlyBinding(binding)) {
+            return false;
+        }
+        return this.lastModifierKey === binding.code && isModifierKeyActive(event, binding.code);
+    }
+
+    /**
+     * Whether the event is a browser/OS shortcut chord the IME must not consume: a
+     * Ctrl/Cmd modifier is held and it is not the toggle key being held (which would
+     * make it the IME key, not a modifier; see isToggleKeyHeld).
+     */
+    private isShortcutChord(event: KeyboardEvent): boolean {
+        return hasShortcutModifier(event) && !this.isToggleKeyHeld(event);
+    }
+
+    private hasNonCollapsedSelection(): boolean {
+        const selection = document.getSelection();
+        return !!selection && selection.rangeCount > 0 && !selection.getRangeAt(0).collapsed;
+    }
+
+    /**
+     * Shift+Backspace: lift the character before the caret back into composition so
+     * its last jamo can be edited. Deletes the character and re-enters composition
+     * with it. Returns false (handling nothing) when the adapter cannot read/delete
+     * the previous character or it is not composable Hangul, so the caller can fall
+     * through. Shared by the capture-phase guard (normal path) and the bubble handler.
+     */
+    private handlePreviousCharComposition(event: KeyboardEvent): boolean {
+        if (!this.compositionAdapter.supportsMethods("getPreviousCharacter", "deleteContentBackwards")) {
+            return false;
+        }
+
+        const character = this.compositionAdapter.getPreviousCharacter();
+        if (!this.compositor.setCharacter(character)) {
+            return false;
+        }
+
+        this.compositionAdapter.deleteContentBackwards();
+        this.compositionAdapter.beginComposition(character!, KeyCode.Backspace);
+
+        cancelEvent(event);
+        return true;
+    }
+
+    /**
+     * Process a jamo key: add it to the in-progress block and cancel the key so the
+     * editor does not also act on it. Shared by the capture-phase guard (for the
+     * composition-starting jamo over a selection) and the bubble-phase handler.
+     */
+    private handleJamoKey(event: KeyboardEvent): void {
+        const code = event.code as KeyCode;
+        const key = keyMap[code];
+        if (!key?.jamo) {
+            return;
+        }
+
+        const jamo = event.shiftKey && key.jamo.shift ? key.jamo.shift : key.jamo.normal;
+        this.addJamo(jamo, code);
+
+        cancelEvent(event);
+    }
+
+    /**
+     * Backspace pressed mid-composition: drop the last jamo and re-render the now
+     * shorter block, cancelling the key so the editor does not also act on it. If the
+     * block is now empty, fall back to the long-standing contenteditable hack:
+     * commit a throwaway "x" and let the editor's own Backspace delete it (so we
+     * deliberately do NOT cancel the key in that case).
+     *
+     * Called from both the capture-phase guard (the normal path) and the bubble-phase
+     * handler (fallback for detached elements); whichever runs first handles it, and
+     * the other is skipped because the compositor is no longer compositing.
+     */
+    private handleComposingBackspace(event: KeyboardEvent): void {
+        const block = this.compositor.removeLastJamo();
+        if (block) {
+            this.compositionAdapter.updateComposition(block, KeyCode.Backspace);
+            this.notifyOnEntry();
+
+            cancelEvent(event);
+        } else {
+            this.compositionAdapter.endComposition("x");
+        }
+    }
+}
+
+/**
+ * Whether a key event carries a shortcut modifier, meaning it is a browser/OS
+ * shortcut (copy, paste, select-all, etc.) rather than text input, so the IME must
+ * step aside and let the browser handle it. `ctrlKey` covers Ctrl shortcuts
+ * (Windows/Linux); `metaKey` covers Command shortcuts (macOS). A modifier+letter
+ * chord is never an intended jamo on any platform, so both bypass composition.
+ */
+function hasShortcutModifier(event: KeyboardEvent): boolean {
+    return event.ctrlKey || event.metaKey;
+}
