+++ src/composition/hanja/hanja-converter.ts
@@ -1,41 +1,18 @@
 import { KeyCode } from "../../keyboard/korean-keyboard-map";
 import { CompositionAdapter } from "../composition-adapters/composition-adapter";
-import { HangulCompositor } from "../hangul-compositor";
 import { HanjaCandidate } from "./hanja-candidate";
 
-export type HanjaConversionTarget =
-    | {
-          kind: "composition";
-          reading: string;
-      }
-    | {
-          kind: "previous-character";
-          reading: string;
-      };
+export type HanjaConversionTarget = {
+    kind: "previous-character";
+    reading: string;
+};
 
 /**
  * Hanja conversion — the distinct phase that turns an already-composed Hangul
- * syllable into Hanja. It deliberately sits *alongside* the compositor rather than
- * inside it: jamo assembly is already done by the time this runs.
- *
- * Two entry points, mirroring how a real IME's Hanja key behaves:
- *
- *   1. mid-composition — the block (e.g. 한) is still being composed; or
- *   2. after a committed syllable — the caret sits immediately after it.
+ * syllable into Hanja. The key listener commits any in-progress Hangul block before
+ * this runs, so conversion only ever reads the committed character before the caret.
  */
-export function getHanjaConversionTarget(
-    compositor: HangulCompositor,
-    adapter: CompositionAdapter
-): HanjaConversionTarget | undefined {
-    if (compositor.isCompositing()) {
-        const reading = compositor.getCurrentChar();
-        if (!isSingleHangulSyllable(reading)) {
-            return undefined;
-        }
-
-        return { kind: "composition", reading };
-    }
-
+export function getHanjaConversionTarget(adapter: CompositionAdapter): HanjaConversionTarget | undefined {
     if (!adapter.supportsMethods("getPreviousCharacter", "deleteContentBackwards", "inputCharacter")) {
         return undefined;
     }
@@ -48,19 +25,7 @@ export function getHanjaConversionTarget(
     return isSingleHangulSyllable(reading) ? { kind: "previous-character", reading } : undefined;
 }
 
-export function commitHanjaCandidate(
-    target: HanjaConversionTarget,
-    candidate: HanjaCandidate,
-    compositor: HangulCompositor,
-    adapter: CompositionAdapter,
-    keyCode: KeyCode
-): void {
-    if (target.kind === "composition") {
-        adapter.endComposition(candidate.hanja);
-        compositor.reset();
-        return;
-    }
-
+export function commitHanjaCandidate(candidate: HanjaCandidate, adapter: CompositionAdapter, keyCode: KeyCode): void {
     adapter.deleteContentBackwards();
     adapter.inputCharacter(candidate.hanja, keyCode);
 }
